backend/requirements.txt
bcrypt==4.1.3
blinker==1.8.2
cachelib==0.13.0
cffi==1.16.0
click==8.1.7
cryptography==42.0.8
Deprecated==1.2.14
Flask==3.0.3
Flask-Cors==4.0.1
Flask-Limiter==3.7.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
Flask-User==1.0.2.2
Flask-WTF==1.2.1
greenlet==3.0.3
importlib_metadata==8.0.0
importlib_resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
limits==3.13.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
msgspec==0.18.6
mysqlclient==2.2.4
ordered-set==4.1.0
packaging==24.1
passlib==1.7.4
pycparser==2.22
Pygments==2.18.0
python-dotenv==1.0.1
rich==13.7.1
SQLAlchemy==2.0.31
tinycss2==1.3.0
typing_extensions==4.12.2
Werkzeug==3.0.3
wrapt==1.16.0
WTForms==3.1.2
zipp==3.19.2
